20201203 조병근
. 내 연차이력 (수정 중)

+++ Circle/src/main/java/com/kh/circle/vacation/entity/Vacation.java
@@ -20,6 +20,6 @@ public class Vacation {
 	private String anva_hstr_sign_code;		// 결재코드 - 연결
 	
 	//calc Values
-	private int days;						// n일
+	private double days;						// n일
 	private int nights;						// n박
 }
